api/requirements.txt
fastapi
uvicorn
psycopg2-binary
prometheus-client
prometheus-fastapi-instrumentator